server/requirements.txt
flask
flask-restfull